+++ internal/provider/provider.go
@@ -165,6 +165,7 @@ func (p *LangSmithProvider) Resources(ctx context.Context) []func() resource.Res
 		NewSCIMTokenResource,
 		NewFeedbackIngestTokenResource,
 		NewPlatformFeatureResource,
+		NewEvaluatorResource,
 	}
 }
 
@@ -194,6 +195,7 @@ func (p *LangSmithProvider) DataSources(ctx context.Context) []func() datasource
 		NewFeedbackIngestTokensDataSource,
 		NewAuditLogsDataSource,
 		NewPlatformFeaturesDataSource,
+		NewEvaluatorDataSource,
 	}
 }